Add resilient-community-apps as submodule
Adding the community apps repo as a submodule so it can be mounted into containers for interactive use.  It can also be used to copy a given integration into a docker image using Dockerfiles which will be bit quicker than cloning the repo into the containers at build time and will also give us smaller images.